fix build errors, document referencing code examples and add comments

+++ documentation-website/Writerside/snippets/exposed-dao/src/main/kotlin/org/example/examples/DeleteExamples.kt
@@ -2,16 +2,12 @@ package org.example.examples
 
 import org.example.entities.StarWarsFilmEntity
 
-const val SEQUEL_ID = 8
-
 class DeleteExamples {
     fun deleteFilm() {
-        val movie = StarWarsFilmEntity.new {
-            name = "The Last Jedi"
-            sequelId = SEQUEL_ID
-            director = "Rian Johnson"
+        val movie = StarWarsFilmEntity.findById(2)
+        if (movie != null) {
+            val deletedMovie = movie.delete()
+            println(deletedMovie)
         }
-        val deletedMovie = movie.delete()
-        println(deletedMovie)
     }
 }
